# openapi/intake-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intake Request API",
    "description": "Creates and submits tenant-isolated innovation intake requests. Operation IDs are stable for Azure API Management MCP projection.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/intake-requests": {
      "post": {
        "tags": [
          "intake requests"
        ],
        "summary": "Create an intake request draft",
        "operationId": "create_intake_request",
        "security": [
          {
            "MicrosoftEntraBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid intake request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Request state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "412": {
            "description": "ETag precondition failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Persistence request throttled",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Persistence operation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "intake requests"
        ],
        "summary": "List authorized intake requests",
        "operationId": "list_intake_requests",
        "security": [
          {
            "MicrosoftEntraBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "continuationToken",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Continuationtoken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeRecordPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid intake request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Request state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "412": {
            "description": "ETag precondition failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Persistence request throttled",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Persistence operation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/intake-requests/{request_id}/submit": {
      "post": {
        "tags": [
          "intake requests"
        ],
        "summary": "Submit an intake request",
        "operationId": "submit_intake_request",
        "security": [
          {
            "MicrosoftEntraBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid intake request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Request state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "412": {
            "description": "ETag precondition failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "428": {
            "description": "If-Match is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Persistence request throttled",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Persistence operation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "schemaVersion": {
            "type": "string",
            "title": "Schemaversion"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "submittedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submittedat"
          },
          "intake": {
            "additionalProperties": true,
            "type": "object",
            "title": "Intake"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "schemaVersion",
          "createdAt",
          "updatedAt",
          "intake"
        ],
        "title": "IntakeRecordResponse"
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "draft",
          "submitted"
        ],
        "title": "RequestStatus"
      },
...
      "IntakeRequest": {
        "title": "Innovation and AI use case intake request",
        "description": "Captures the core idea and optional discovery details for an innovation or AI use case request.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "problemOpportunity",
          "proposedIdea",
          "expectedOutcome",
          "requester"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "A short, recognizable name for the request.",
            "minLength": 3,
            "maxLength": 150
          },
          "problemOpportunity": {
            "type": "string",
            "description": "The current problem, unmet need, or opportunity and why it matters.",
            "minLength": 20,
            "maxLength": 5000
          },
          "proposedIdea": {
            "type": "string",
            "description": "The proposed innovation or AI-enabled approach.",
            "minLength": 20,
            "maxLength": 5000
          },
          "expectedOutcome": {
            "type": "string",
            "description": "The desired result and the change expected if the idea succeeds.",
            "minLength": 10,
            "maxLength": 3000
          },
          "requester": {
            "$ref": "#/components/schemas/IntakeRequest_requiredPerson",
            "description": "The person submitting the request."
          },
          "businessContext": {
            "$ref": "#/components/schemas/IntakeRequest_businessContext"
          },
          "targetUsers": {
            "$ref": "#/components/schemas/IntakeRequest_targetUsers"
          },
          "valueAndSuccess": {
            "$ref": "#/components/schemas/IntakeRequest_valueAndSuccess"
          },
          "aiConsiderations": {
            "$ref": "#/components/schemas/IntakeRequest_aiConsiderations"
          },
          "dataConsiderations": {
            "$ref": "#/components/schemas/IntakeRequest_dataConsiderations"
          },
          "dependencies": {
            "type": "array",
            "description": "Systems, teams, initiatives, approvals, or other dependencies.",
            "items": {
              "$ref": "#/components/schemas/IntakeRequest_dependency"
            },
            "maxItems": 50,
            "uniqueItems": true
          },
          "deliveryConsiderations": {
            "$ref": "#/components/schemas/IntakeRequest_deliveryConsiderations"
          },
          "risks": {
            "type": "array",
            "description": "Known delivery, adoption, operational, legal, or technical risks.",
            "items": {
              "$ref": "#/components/schemas/IntakeRequest_risk"
            },
            "maxItems": 50
          },
          "responsibleAI": {
            "$ref": "#/components/schemas/IntakeRequest_responsibleAI"
          },
          "securityAndPrivacy": {
            "$ref": "#/components/schemas/IntakeRequest_securityAndPrivacy"
          },
          "ownership": {
            "$ref": "#/components/schemas/IntakeRequest_ownership"
          },
          "supportingLinks": {
            "type": "array",
            "description": "Links to research, process documents, prototypes, or other evidence.",
            "items": {
              "$ref": "#/components/schemas/IntakeRequest_supportingLink"
            },
            "maxItems": 20
          },
          "additionalNotes": {
            "type": "string",
            "description": "Relevant information not covered by the structured fields.",
            "maxLength": 5000
          }
        }
      },
...
    },
    "securitySchemes": {
      "MicrosoftEntraBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token."
      }
    }
  }
}